Move totalChl computation into MARBL
Additional surface_forcing_output field for totalChl, and computation is done
in marbl_set_surface_forcing instead of relying on the driver to figure it out.

The next step will be to generalize the surface_forcing_output datatype
(similar to diagnostics, forcing fields, and whatnot), rather than hardcoding
the specific 'flux_o2', 'flux_co2', and 'totalChl' fields. Then the driver can
request which (if any) of the fields it wants.