Generalize surface_forcing_output_type
Much like forcing fields and diagnostics, there is now a single_sfo_type that
contains both data and metadata for a single surface forcing output field as
well as a wrapper marbl_surface_forcing_output_type class that contains an
array of the single types. Because the number of surface forcing outputs are so
small, and because the driver needs to initialize them rather than having it
set up in marbl_init, I use an allocatable array of the single sfo type and
every time a new field is added we allocate a new array of size (n+1), copy new
data over, and then deallocate the old data. If the number of potential fields
grows, we will want to rethink this strategy.