First pass removing 'error code returned' messages
I created a new subroutine to standardize the error traceback in MARBL. When
first encountering an error, call LOG%log_error() and return; every subsequent
call should be to LOG%log_error_trace, which takes the name of the routine that
produced the error as an argument rather than a log message.

Note that this is just a wrapper script - log_error_trace creates the log
message

"Error reported from [routine name]"

and then calls LOG%log_error with that message.

I say "first pass" because I still need to adapt the errors from
add_diagnostics() and add_forcing_field(), where I want to include the name of
the diagnostic [or forcing] field that produced the error. I was originally
thinking that I should add an optional argument for field name, but instead I
might treat add_forcing_field([FIELDNAME]) as the routine name.