Add description of logging module to source code
This commit only affects the comments in .F90 files, but there is now a header
in marbl_logging.F90 that explains how marbl_instance%StatusLog works and how
the GCM should interact with it. Also, there was a "TODO" comment in
marbl_interface_types.F90 that had already been done, so I deleted it.