Don't reorder tracers; add living tracers to class
Instead of reordering the tracers so that all the non-living tracers come first
and all the living tracers follow, keep the tracers sorted by module but
include living tracer indices in the marbl_tracer_index_type class. This commit
does NOT run -- I need to move marbl_tracer_indices into marbl_interface_class
so that the constructor is called after the autotrophs have been initialized.
On the other hand, it's Friday afternoon and I want this code to be accessible
during the Spring Powerdown this weekend (otherwise I would not commit broken
code).