Trying to move more metadata from driver to MARBL
In order for the driver to be ecosys_module-agnostic (i.e. to not care if
tracer comes from general ecosys or CISO), we need more initialization-based
metadata available in MARBL. To that end, I added a marbl_tracer_read_type
array to the marbl_instance class and values are populated after the namelists
are read. I also added a tracer_module_name field to the tracer metadata so the
driver can know (on a per-tracer basis) whether a tracer is ecosys or ciso; the
driver just does not know the range of indices corresponding to the two or how
many tracers are in each module. (For now the driver ignores these fields but
thatwill change next week.)

Other improvements clean up how we track the number of tracers / which module
they belong in. For starters, ecosys_used_tracer_cnt is now determined in MARBL
rather than the driver (and the GCM passes in what it expects the MARBL tracer
count to be for comparison); ecosys_ind_beg, ecosys_ind_end,
ecosys_ciso_ind_beg, and ecosys_ciso_ind_end are all determined in MARBL as
well (as part of the constructor for the tracer indices).