Remove num_tracers from marbl_ciso interfaces
marbl_ciso_set_interior and marbl_ciso_set_surface_forcing both required
num_tracers as an intent(in), but that variable is only used to provide
dimensions for other arguments to the routines. I replaced those uses with a
dimension of ':'.

I also cleaned up an if statement in the marbl_tracer_indexing_type constructor
- if an if statement has a return in it, then 'else' is un-necessary. This is
paving the way to moving a couple of error checks out of marbl_ciso and
possibly marbl_mod (we can compare the number of CISO tracers to
ciso_tracer_cnt in the constructor, rather than while initializing tracer
metadata)