Remove duplicate code in marbl_mod.F90
marbl_tracer_update_file_metadata() ran through basically the same code twice,
once to process tracer_init_ext and once to process ciso_tracer_init_ext. To
remove this duplicate code, I updated the interface and now we call the routine
twice, once with tracer_init_ext as an intent(in) and once with the ciso copy
instead. (So marbl_interface calls the routine twice instead of once, and the
routine runs once instead of twice.)