Add rank to saved state type
Currently only support rank 2 and rank 3 saved state variables, and rank has to
be compatible with vgrid ('none' for rank=2, and either 'layer_avg' or
'layer_iface' for rank=3). I hope this allows us to add rank=0 for global sums,
but I'm not sure exactly what that will look like since MARBL will need to pass
out the full 2D field so the GCM can do the area-weighting for computation.