Driver from enhancements/write_full_namelist
The updated initialization will provide a better way to log all the namelist
parameters than we had been working on in an older branch. This commit brings
updates to the library Makefile as well as introduces a stand-alone driver that
can be used to run simple tests. Currently the only included test runs
marbl_instance%init and prints the resulting marbl_instance%StatusLog -- the
goal is to include all parameter settings in this output. On the old branch we
were going to do that by writing a namelist to the log, but with a new
parameter datatype it makes more sense to roll our own write routine.