Cray compatibility issues
Updated driver makefile to support Cray, realized that ftn is a little
finicky in how it reads text files... so the read of 'marbl_in' is less
elegant than it used to be. Also added &marbl_ciso_nml to marbl_in and
set up init_output.sh to handle non-gnu compilers.