Clearer logic in linear_root()
To find a root between x(1) and x(2), we require either y(1) and y(2) to be
opposite signs OR one of the two values to be 0. To check for a sign change, we
originally look to see if y(1)*y(2) > 0... but in this commit we update that to
explicitly look to see if (y(1) > 0 && y(2) > 0) or (y(1) < 0 && y(2) < 0).
This offers two improvements over the old code:

1) more detailed error message (can now say "both y values are positive" or
   "both y values are negative" instead of "y values have same sign")
2) Avoid potential for underflow (multiplying two very small positive numbers
   could result in 0)