Clean up dependency generator
Dependency generator for the stand-alone build does not include duplicate lines
(previous version generated a .d file wth one line per use statement in the
code, this version prints out one line per dependency). Not strictly necessary
for this branch, but it's a nice improvement to have and I think we'll be using
the stand-alone build more frequently now that the driver has been added.