Only allocate memory in marbl_parms once
There are allocatable arrays that are read in by ecosys_restore_nml; when the
arrays were populating local variables, it was fine to allocate memory every
time. Now we need to check to see if the module variable in marbl_parms has
been allocated to avoid calling allocate a second time.

I think we will want to completely overhaul this aspect of initialization for
tracer restoring when we make init thread-safe. (Likely we will need to have
each marbl_instance contain its own marbl_parms derived type, so they will all
be allocated.)