Remove marbl_single_parms_define()
The marbl_parms_type constructor had been passing the parameter short name to
add_parms(), which called the single_parms_define() call, which had a select
case ('sname') to populate the rest of the metadata. Now the constructor passes
sname, lname, units, datatype, and the correct pointer to add_parms(), which
updates the single_parms_type and removes the need for the single_parms_define
routine. The datatype is checked to ensure that you pass the correct pointer,
otherwise MARBL aborts.