Move marbl_parameters to marbl_interface_type
Instead of a module variable in marbl_parms.F90, we now have
marbl_instance%parameters so each instance carries the parameter metadata.
There is a new check in marbl_parms_construct to make sure the constructor is
only called once (which won't be a problem now that it is part of the MARBL
interface).

Also, dropped the second 'a' in 'params' from marbl_params_init and
marbl_params_set_defaults (they are now marbl_parms_*).