Introduced marbl_config_mod
Goal for this commit was to move marbl_restore_nml (both setting default values
and reading the namelist) into marbl_parms.F90. This was a little more
difficult than anticipated because we need to know the tracer count before
setting up the arrays in marbl_restore_nml (the arrays determine which tracers
we apply restoring to). For now, the fix is to move autotrophs and zooplankton
to marbl_config_nml, and then initializing in the following order:

1. set defaults / read marbl_config_nml
2. determine tracer count
3. set defaults / read all other namelists (in marbl_parms.F90)

I still need to introduce the marbl_instance%config() routine in the interface,
and need to make the namelist arguments optional. Then I need to write the put
/ get routines and make sure I lock the config put in the call to init. This
will require a separate marbl_config_type that looks similar to
marbl_parms_type (but I think the constructor will be different for the two).