Add marbl_instance%config routine
Moved some of the setup from marbl_instance%init to a new configuration
routine. For now, config just sets ciso_on and reads marbl_config_nml but in
the future more "initialization" will be done in this phase instead.

Also updated the formatting of some output in the log to make output easier to
read (added more whitespace)