Clearer default names for zooplankton / autotrophs
Instead of "unknown" for autotrophs past the third (first three are sp, diat,
and diaz) and zooplankton past the first (which is just zoo), default for the
former is autoN and zooN for the latter. So 5 autotrophs and 2 zooplankton
would be

autotrophs: sp, diat, diaz, auto4, auto5
zooplankton: zoo, zoo2

It may make sense in the future to add some leading zeros auto0004, auto0005).