Set derived config / parms when locking datatype
Some variables in marbl_parms_mod were set based on the value of namelist
variables (for example, caco3_bury_thres_iopt is an integer corresponding to
string values of the namelist variable caco3_bury_thres_opt). To be compatible
with put / get statements, these flags need to be set after the user has a
chance to put() values from the GCM, so these are now set in a separate
routine. Making this change highlighted the fact that marbl_forcing_tmp_nml
needs to be set in config rather than init, because the behavior of init
(setting up the marbl_forcing_field_type) depends on these variables. So
marbl_forcing_tmp_nml is now in marbl_config_mod, and there is a
set_derived_config() routine that is called from init when locking
marbl%configuration.