Add strawman put() for config and parms
In GCM, can use

call marbl_instance%configuration%put(varname, value, LOG)

or

call marbl_instance%parameters%put(varname, value, LOG)

to change config / parm settings. The put command is overloaded, so value can
be r8, integer, logical, or a string. The routine that actually does the put
first checks to ensure the data type is not locked, and returns an error if it
is. Other errors can be triggered if the variable name is not recognized or if
the datatype of value does not match the data type of the parameter you want to
set (you _can_ set real-valued parameters with an integer).