Add strawman get() interface
This may change, but for now GCM can use

call marbl%configuration%get('varname', val, marbl%StatusLog)
call marbl%parameters%get('varname', val, marbl%StatusLog)

Where val is an intent(out) that must match the datatype of 'varname' in the
derived type. Providing the wrong data type (trying to get() a real parameter
into an integer val, for example) will result in an error.

This routine is not used in the driver yet, but it was tested fairly
extensively without committing those tests at this time.