Add check to see if GCM can perform global ops
There is a new optional argument to marbl%config() [lgcm_has_global_ops] that
must be set to true if the MARBL setup requires global operations. For now,
this means that ladjust_bury_coeff = .true.

While this flag is an argument to marbl%config, the check is actually done in
marbl%init after locking the configuration. MARBL will abort if
lgcm_has_global_ops = .false. and ladjust_bury_coeff = .true.