New test to make sure get / put works
Test is simple: every variable is set to either .true. (if logical), -1 (if
integer or real), or '-1' (if string). Then, one by one, we make sure variable
has the set value before changing to .false. or a positive integer. If the put
call for two variables change the same value in memory, the second variable
will fail the test.

Expected output:
```
marbl_in contained 40 characters
Beginning get_put test...
Setting variables to .true. or -1 ...
... Done!
Making sure variables are .true. or -1 then setting to .false. or n ...
... Done!
Setting variables to .true. or -1 ...
... Done!
Making sure variables are .true. or -1 then setting to .false. or n ...
... Done!
```