Improvements to setup output / error checks
Output in the log is now formatted as so:

1. &marbl_config_nml
2. MARBL tracer indices
3. Requested surface forcing fields
4. MARBL parms

There is also a consistency check after step 2 to ensure that the tracer count
matches what was built (if you build with 27 base tracers but set up config in
such a way that you only have 26 tracers, the run will abort in
initialization).

There's also a small tweak to the driver Makefile to remove marbl.exe before
trying to build libmarbl.a, otherwise a test will run with an old executable if
the build fails in the library.