Bug fix for tracer restoring
MARBL provides the GCM with a list of tracers it wants restoring climatologies
for, but with the changes in this branch the GCM also needs its own list of
tracers to restore (and the file(s) containing the climatologies). Moreover,
those two lists do not need to be in the same order. (In fact, every tracer on
the MARBL list must appear on the GCM list but the converse is not true.) The
bug-fix accounts for lists with variables appearing in different orders, as I
had messed up the indices in restore_tracers().

Also cleaned up some log output and removed an unused field from the
marbl_interior_forcing_input_type -- I had added a field for the climatology
data, but that is already available from marbl_restore_type.