request_restoring test works again; bugfixes
Although the 'request_forcing' test prints out the tracers that will be
restored, having the 'request_restoring' test provides an example of the logic
required in the GCM to determine what tracers will be restored... so I kept the
test in and it works again.

Also, there were a handful of log calls that had the message and subname
arguments reversed (3 that I added on this branch, plus two others that I also
fixed).