Need to call set_rank for surface forcing
All surface forcing fields have rank 0 currently, but that was not being
explicitly set. Code ran fine on intel, where the compiler set rank=0 in the
derived type, but gfortran populated rank with unknown values that causing
allocate_memory not to allocate memory at all.

It might be wise to put an error condition in allocate_memory() if rank is not
0 or 1, but currently we don't pass allocate_memory the status log. Maybe a
better option would be to call set_rank from allocate_memory? That would
involve moving where allocate_memory is called for interior forcing fields and
passing rank, [optional dim1], and a status log... at which point maybe we
don't need to save the rank and extent in the metadata?