Add MPI support to stand-alone driver
Run any test with "--mpitask #" and instead of running

$ [MARBL_ROOT]/tests/driver_exe/marbl.exe

The test scripts will execute

$ mpirun -n # [MARBL_ROOT]/tests/driver_exe_marbl

This required updating marbl_testing_class.py to handle the --mpitask option
and also edit the Makefile to build with mpif90 instead of the serial compiler.
Two obvious fixes that need to go in:

1. Handle MPI distributions that don't run via mpirun
2. Handle MPI builds where the compiler is not mpif90

This introduces two new compile macros to the driver build: -DUSE_GPTL and
-DHAVE_MPI; the only library files that use these macros is
marbl_timing_mod.F90.