More timer support: GPTL via POP wrappers
Added support for the CCSMCOUPLED compiler macro, which then has MARBL
use the POP timer module (and some domain information). This is not
ideal, but I don't think I'll be able to directly access the GPTL
library so it's probably the best option.

Note that this adds a new flag to the "add_timer" routine to tell
whether a timer is in a CESM-threaded region. It's an optional argument
and is ignored in all non-CCSMCOUPLED builds.