Add OpenMP directive to check for threading
When summarizing timers, we need to know if marbl_instance(:) ran concurrently
or not; to determine, we see if OpenMP is available and, if so, see if number
of threads > 1.

Also removed the "been_run" flag; a runtime of 0.000s should be treated as such
whether the timer was ever enabled or not.