Remove locmip_k1_k2_bug_fix from config
MARBL should not be run with this variable set to .false. so I removed the
variable and changed the code to only follow the .true. path.

Note that this update also removes the k1_k2_pH_tot variable from
marbl_comp_co3_coeffs() as that variable is also always set to .true.