Add category to marbl_config_and_parms_type
The marbl_single_config_or_parm_type has a category id object; it is an integer
that refers to an index in the category array owned by the
marbl_config_and_parms_type. When you add a new configuration variable or
parameter, you should specify a category and then variables in the same
category will be written to the log together.

Note that this is just the category infrastructure commit, at this point the
category is not being used at all.