More Makefile cleanup
We define $(MAKE_DIR) to be the location of the Makefile, and then set default
values of SRC_DIR, OBJ_DIR, etc relative to MAKE_DIR. This lets you run

$ make -f $MARBL_DIR/src/Makefile

to build libmarbl-gnu.a

(You no longer need to be in the src/ directory when you invoke make).

I also tried to add better comments describing what the Makefile is doing,
because defining $(MAKE_DIR) is non-trivial.