Two updates to marbl log data type
1. Default ElementInd = -1; in practice, ElementInd = -1 => no location was
   given when the message was logged.
2. instead of lall_tasks = .true. if message should be written by all tasks
   (and .false. if only master task should write the message), the variable to
   use is lonly_master_writes. As the name implies, if .true. then the message
   should only be written from the master task and .false. means that every
   task should write the message.

In (2) above, note that "write the message" means either print to stdout or
write directly to a log file; MARBL will provide the log type with the message
to return to the GCM regardless of the setting of this flag, but it's a way to
let the GCM know that it might be recieving the same log message from multiple
tasks. One such example is when MARBL logs the value of all the namelist
variables.