Remove mask from co2calc subroutines
Now pass num_elements (array size) and num_active_elements (number of elements
in the array on which to act); for surface forcing, these are the same number.
For interior forcing, num_elements = km and num_active_elements = kmt.

Note that drtsafe has a local mask variable that is passed to total_alkalinity:
this is masking out which points have converged in the iterative solver to
reduce computation.