Omit comment from log if it makes string too long
There's an issue on PGI (that smells like a compiler bug) where the
actual contents of a string of length char_len are being ignored in
favor of char_len characters of gibberish; this string is being appended
to another string that has length char_len, causing an overflow. The fix
is to ensure that len_trim() of the two strings (plus the three
additional characters being included) don't exceed the length of the
string they are being written into. If the limit is exceeded, a warning
is printed out and the comment is omitted.