Remove marbl_nl_cnt and marbl_nl_buffer_size
These parameters are not needed in the MARBL library; restrictions on
the size of the namelist are entirely up to the GCM and marbl can use
the len() instrinsic on intent(in) string arguments to determine what
the max length is.

This affects several files because the interface, configuration module,
and parameter module were all using these parameters (un-necessarily).

The driver for stand-alone testing now defines these parameters itself
(as any GCM will need to if using namelists to change MARBL config
variables and parameters.)