Introduced marbl%init subroutine
If you are initializing via namelist and don't want to use marbl%put() or
marbl%get(), this routine runs all three stages of initialization in a single
call.

The init_from_namelist test uses it (so interface is different in the "no
namelist" test), and I also updated the stand-alone test suites so the MARBL
log ontains the subroutine names updated in the last commit rather than the old
subroutine names.