Rename marbl_total_tracer_cnt
In the tracer indexing type, the component is now simply total_cnt;
elsewhere in the code, the variable is tracer_cnt. This discrepancy may
seem a little odd on face value, but the "tracer" modifier doesn't seem
necessary in the tracer index type (what else would we be counting?) and
the "total" isn't necessary elsewhere in the code (whereas in the type
we also have counts for individual tracer modules).