Update how put() calls work
Now call put() before initializing code, and it is applied automatically when
the variable is added to the parameter type. Note that put calls are now made
before we initialize marbl_instance%StatusLog, so a string is returned (empty
=> no error in put call).

Still to do: combine configuration and parameters into one type (will require
moving where we check to see that user_settings is empty).