Update how "put()" works for parameters
Instead of calling marbl%configuration%put() or marbl%parameters%put(), GCMs
now call marbl%put() [currently with a string argument detailing which phase
the variable is defined in -- configuration => 'phase2' and parameters =>
'phase3'; that argument will be removed once all variables are defined in
marbl%settings].

This setup allows us to call marbl%StatusLog%construct from the put call, so we
can report errors from put() even after moving it prior to the init() call. The
next commit will reinstate the error check to make sure put() is not called
after init.

To make this possible, the marbl_logging_type was updated to include a private
logical variable (lconstructed) which is .false. originally but set to .true.
in the constructor. The constructor immediately returns if it is called after
lconstructed is changed to .true.