First step to removing namelist
Completely removing the "group" data from marbl_single_setting_ll_type will be
very invasive (because a local variable is set just prior to every add_var call
in marbl_parms.F90), so for now I left the component of the type in place but
the log output does not check the group data anymore.

Also, I fixed a dependency issue in the stand-alone driver Makefile: if
libmarbl.a changes, then all the driver object files need to be rebuilt.