Add array-handling to put_all_string()
For now this can only handle arrays of strings and reals, integers and logicals
will have to come later. Note that I added improved error-handling because

parm_scalelen_vals = 1.2, 2.3, 3.4, 4.5

was being treated the same as

parm_scalelen_vals = 1.2

And we don't want that.

Before adding support for logical and integer arrays, I need to refactor
marbl_interface.F90 so the string processing is only done in one place. Right
now, for example, we check to see if a string contains a real number in 3
different places.