Better thread safety on new module variables
Now that autotrophs_config, zooplankton_config, etc are allocatable module
variables, we need to ensure that they are only allocated once when running
with multiple threads. At this point we still require initialization to be run
serially, but we now check that memory is not allocated before allocating it
(and also check that memory is allocated before deallocating it at shutdown()).
There was also a bug in the diagnostics relating to memory being allocated but
not deallocated.