'local' machine checks $PATH for compilers
Instead of having local-gnu and local-pgi (and needing to add local-intel or
local-nag if requested), there is a single 'local' machine that checks to see
what compilers are in $PATH to determine supported_compilers.

The search is done in the following order:

gnu
pgi
intel
nag

The first one found is the default, and if none are found then an error is
reported.