Update init() routine to reduce footprint
Two classes of parameters are initialized back-to-back: the general parameters,
and the PFT setup (short names, etc for autotrophs, zooplankton, and grazers).
Rather than have two separate calls from marbl_interface, there is a single
routine in marbl_init_mod that sets defaults and defines the parameters for
both classes in the correct order.