Better array parsing in puts; clearer comments
Much like with namelist,

varname = value

Is the same as varname(1) = value for an array. Did not extend to 2D arrays or
beyond.

Also cleaned up the repeated comments refering to the note about marbl_parms
framework (the note itself is now where module variables are declared, and the
comment refering to it mentions marbl_settings framework rather than
marbl_parms framework)