improve unit testing
Added a few more linear root tests (included two that expect to not find root),
as well as a few more str_to_array tests. Also created strip_comments_test that
mimics what we do when parsing input file lines -- uses '!' instead of ',' as a
delimiter and makes sure that the first element of array is returned correctly.