Add get_tracer_cnt routine to interface and tests
The get_tracer_cnt test returns the number of tracers in the ecosys_base
module, the number of tracers in the ciso module, and the total MARBL tracer
cnt all in one line, separated by spaces. So with ciso_on = .true. and all
other parameters kept to default settings, the output of the test is

Beginning tracer_cnt test...
32 14 46

One use for this test will be for POP to determine ECOSYS_BASE_NT, CISO_NT, and
MARBL_NT directly from the MARBL code rather than relying on users to update it
correctly.