Make YAML defaults consistent with Fortran
the autotroph(:)%Qp_fixed parameter is computed in Fortran as either
1/117 on 0.32*(1/117) [parentheses due to using a parameter set to
1/117]; in double precision, 0.32/117 != 0.32*(1/117). Also, the python
code was only printing 16 significant digits, when 17 are needed to
ensure bit-for-bit between the decimal and expression.

Given that last point (17 sig digits), I updated all log output in the
Fortran code  to be E25.17 instead of E24.16 -- since Fortran uses
0.1e+1 instead of 1.0e0 to represent 1, the digit in the ones place is
not significant (in the Python code, we use %24.16e because the ones
place is also significant).