Previous fix to restore tracers ignored CISO
This update makes sure interior restoring is applied to ALL tracers; it
is done in a manner such that restoring is applied after the budget
computations, so those no longer need to be adjusted to subtract out the
restoring term.

Also cleaned up one change from previous code, reverting back to
convention of putting mathematical operator after the continuation
character when wrapping lines.