Change how python scripts determine tracer count
default_settings.json now contains _tracer_list instead of _tracer_cnt; it is a
dictionary that is used in conjunction with the rest of settings_dict to
determine which tracers are being requested. get_tracer_count now returns the
length of the list of all requested tracers (and there is a get_tracer_list
routine to provide the tracers themselves as well).