Better template structure in diagnostics.yaml
Based on discussions at Jan 23rd MARBL meeting, a lot of re-writing code that
processes default_diagnostics.json. Some highlights:

1. instead of a 'module' key, an optional 'dependencies' key is used to specify
diagnostics that are not always defined
2. Better use of templating for per-tracer diagnostics (still need to expand to
PFTs)

The process is now:

1. default_diagnostics.json is read into self._diagnostics
2. self._diagnostics is "processed" into self._resolved_diagnostics:
   i.   diagnostics that are not available are removed
   ii.  all templates are filled
   iii. if 'frequency' is a dict, proper value is determined
   iv.  frequency / operator are converted to lists even if just single value
3. self.diagnostics_dict is determined from self._resolved_diagnostics
-- I don't know if we really want this to be part of the diagnostics class, or
if this should really be part of the generate_diagnostics script. If the
latter, I will rename self._resolved_diagnostics to self.diagnostics or
something similar.