Adding more diagnostics to yaml / json
To support some autotroph metadata, I reworked how some of the
dependency logic was determined. determined. "template_fill_dict" is
turning out to be extremely useful, as it can now play a role in
determining the diagnostic name (and values of some diagnostic metadata)
as well as in the dependency check and determining the proper frequency
of diagnostic output.