Multi-instance support & placeholder for new test
We want to introduce a new test to run set_surface_forcing() and
set_interior_forcing(), ideally on multiple columns of data. So marbl_instance
is now marbl_instances(:) in the driver, though the existing tests require
num_inst == 1. There is a new marbl_set_forcing_drv.F90 and associated
tests/regression_tests/set_forcing/ directory but the test is not ready yet.