Better Makefile support for netCDF and more
To build with netCDF, run

$ make NETCDF_INC="-D_NETCDF -I[inc_dir]" NETCDF_LNK="-L[lnk_dir] -lnetcdf -lnetcdff"

I still need to work that into the python scripts with a '--netcdf' option
(first I'll need to figure out best way to find [inc_dir] and [lnk_dir])

Also, I have a plan for handling netCDF calls: public interface will allow
users to specify either a file name or a file identifier, which will be used to
find the marbl_file_entry object in the file_database linked list... that entry
will be passed to a private subroutine which pulls the file identifier out of
the object. This has been implemented for marbl_io_close(), an interface to
marbl_io_close_by_id() and marbl_io_close_by_name(); both of those routines
call marbl_io_close_by_database(), which is private.

This helps in error-checking: if an id or name is not found, then the
by_database() routine will not be called and the public routine will return
with an error. It's also more efficient than the previous implementation, where
the by_name() routine called the by_id() routine and we searched through
file_database twice to find the same match.