zoo1[34]Ctot -> zootot1[34]C
Part of some post-review code-cleanup - realized that the
zooplankton_share_type and marbl_ciso_mod.F90 should be using zootot instead of
zoo to signify this is a sum over zooplankton species. As a first pass, I
changed zoo13Ctot and zoo14Ctot to zootot13C and zootot14C, respectively.