Streamline how we set up python environment
I forked the rtfd/sphinx_rtd_theme repository into the marbl-ecosys
organization and made version-dropdown the default branch; py_requirements.txt
now points to this branch on github. This makes it easier for users to create
the same sphinx environment that we use to generate the docs, and also lets
travis CI use the same theme version.