More type and variable renames
Biggest changes are

* autotrophs -> autotroph_settings
* zooplankton -> zooplankton_settings
* grazing -> grazer_settings

Those three require YAML (and JSON) changes, and then python
modifications due to the changes in the JSON. Plus updates to the POP
python wrappers.

This commit is still bit-for-bit in POP testing, but all tests will now
fail the namelist comparison due to changes in the variable names.