Building with netCDF is now optional
marbl_testing_class provides a '--netcdf' option, and tests can specify that
netcdf is needed by passing RequireNetCDF=True to parse_args(). This determines
whether the Makefile receives USE_NETCDF=TRUE or not.

Lots of updates to marbl_io_mod.F90 to build correctly when -D_NETCDF is not
specified.