2inst and 5inst history files should be identical
Once baselines have been generated, comparisons to those baselines need to
account for round-off differences between different machines... but the
2-instance and 5-instance tests run on the same machine should be bit-for-bit
so we set atol and rtol to 0. Also improved the error message encountered when
a value in the baseline is 0 and the corresponding value in the new file is
non-zero.