Timers now work with compute_cols test
I needed to add a marbl_instance%shutdown() call to the compute_cols test
subroutine. This highlighted an issue in marbl_interface.F90 where multiple
instances might be sharing the tracer_restore_vars array, in which case it only
needs to be deallocated once. This may actually be a bug -- do we want to
support two instances that apply restoring to different sets of tracers? If so,
then currently those two instances can not be run on the same MPI task because
there is only one tracer_restore_vars array per MPI task.

I also modified summarize_timers to account for tests that run with multiple
instances (for now, the timers just come from instance 1, but future work can
combine timers in a smart way).

For testing, I think it makes sense to run with ninst = 1 and then compare the
results from ninst = 2 to make sure that dividing columns among instances
doesn't change answers; I've kept the 5 instance namelist for now in case we
want to use it later. I can see an argument for testing 1 column per instance
(i.e. ninst = 5) against 1 instance for all columns, in which case maybe we
would drop the ninst = 2 namelist?