Make temp_func_form_opt and Tref per-PFT settings
Also, introduce the temp_func_form_opt = 'power' option (which will be
used by coccolithophores). Note that this commit changes the default
Tref behavior -- rather than depending on temp_func_form_opt (Tref = 30
for q_10 and Tref = 25 for arrhenius), the default is currently based on
the PFT shortname and set to 30 for all current PFTs. I have some ideas
on how to change the way we apply dependencies to allow Tref to depend
on the per-autotroph temp_func_form_opt but haven't implemented that
yet.