Update comments in marbl_pft_mod
Some default variable settings were missing the comment to see the CESM
NOTE in marbl_settings_mod, others had an incorrect message (claiming
the CESM NOTE was in marbl_pft_mod rather than settings_mod).

Also, changed the order in which the autotroph settings are set so that
it matches the order variables appear in the settings YAML file.