Clean up comments and variable names
Following the refactor to move Tref and temp_func_form_iopt onto the autotroph
and zooplankton settings types, the variable Tfunc_ind was no longer an
appropriate name (it wasn't just indexing Tfunc). So Tfunc_ind is now PFT_ind
(matching the "PFT_cnt" dimension in the comments). I also cleaned up the
comments about Tref in the function header to note that the setting is not used
when Tfunc is computed with the power function.